[bundle-size-checker] Extract findExportedPaths into its own module
